requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.2
python-dotenv==1.0.0
openai==0.28